+++ src/lib.rs
@@ -64,6 +64,7 @@ pub use crate::algorithms::operators::disjoint_union::{disjoint_union, disjoint_
 pub use crate::algorithms::operators::intersection::intersection;
 pub use crate::algorithms::operators::simplify::simplify;
 pub use crate::algorithms::operators::union::union;
+pub use crate::algorithms::paths::astar::a_star_path;
 pub use crate::algorithms::paths::dijkstra::{
     DijkstraAllPaths, DijkstraMode, DijkstraPaths, dijkstra_all_shortest_paths, dijkstra_distances,
     dijkstra_distances_cutoff, dijkstra_distances_cutoff_with_mode, dijkstra_distances_multi,